Major Update per inoltro e ricezione dei messaggi
messageCenter.jsp è la pagina dell’admin per poter gestire i messaggi
degli utenti che chiedono spiegazioni

[MessageListAdapter, MessageAdapter, LocalDateTimeAdapter] sono 3 classi
per trasformare le liste di messaggi in formato JSON

Message, DAOMessage e MessageManager sono le classi per la gestione su
DB dei messaggi

NUOVE SERVLET: RetrieveMessage, SendMessage, UploadServlet (per salvare
le immagini) GetImages (per prenderle dalle cartelle)

- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -

SISTEMATO:
itemEntry.jsp ora permette il caricamento delle fotografie (sistemare la
loro visualizzazione)

DAOItem: mancava il caricamento dell’attributo type dal metodo
getFromResultSet per gli Item.

Item: modificato a float il peso degli articoli per problemi di tipi.

AddProductServlet modificato un pochino.

+++ NibInk/src/main/java/com/model/Item.java
@@ -8,13 +8,13 @@ public class Item
 	private String color;
 	private float price;
 	private float vat;
-	private int weight;
+	private float weight;
 	private String dimensions;
 	private String description;
 	 
 	public Item() {}
 	
-	public Item(String codenumber, String title, String type, String color, String dimensions, String description, float price, float vat, int weight)
+	public Item(String codenumber, String title, String type, String color, String dimensions, String description, float price, float vat, float weight)
 	{
 		this.setCodenumber(codenumber);
 		this.setTitle(title);
@@ -52,7 +52,7 @@ public class Item
 		this.vat = vat;
 	}
 	
-	public void setWeight(int weight) {
+	public void setWeight(float weight) {
 		this.weight = weight;
 	}
 	
@@ -89,7 +89,7 @@ public class Item
 		return vat;
 	}
 
-	public int getWeight() {
+	public float getWeight() {
 		return weight;
 	}
